requirements.txt
accelerate==1.9.0
datasets==4.0.0
decoder==0.5
numpy==1.24.3
openai==1.97.1
pandas==1.5.3
peft==0.16.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.4
scikit_learn==1.7.1
sentence_transformers==3.3.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
transformers==4.53.2
